# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Meal {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  healthScore: number;
  cuisine: string;
  tags: string[];
  price: number;
  isAiRecommended: boolean;
  /** @nullable */
  restaurantId?: number | null;
}

export interface MealPlanDay {
  day: string;
  breakfast: Meal;
  lunch: Meal;
  dinner: Meal;
  totalCalories: number;
  totalProtein: number;
}

export interface MealPlan {
  weekOf: string;
  days: MealPlanDay[];
}

export interface Restaurant {
  id: number;
  name: string;
  cuisine: string;
  rating: number;
  deliveryTime: string;
  /** @nullable */
  imageUrl?: string | null;
  isHealthy: boolean;
  tags: string[];
}

export interface GroceryItem {
  id: number;
  name: string;
  category: string;
  quantity: string;
  unit: string;
  isChecked: boolean;
  /** @nullable */
  nutritionNote?: string | null;
}

export interface GroceryList {
  id: number;
  weekOf: string;
  items: GroceryItem[];
  totalItems: number;
  checkedItems: number;
}

export interface GroceryPlanInput {
  goal: string;
  dietaryPreferences?: string[];
  /** @nullable */
  budget?: number | null;
}

export interface UserProfile {
  id: number;
  name: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  weight?: number | null;
  /** @nullable */
  height?: number | null;
  goal: string;
  dietaryPreferences: string[];
  wellnessScore: number;
  streak: number;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface UserProfileInput {
  name?: string;
  age?: number;
  weight?: number;
  height?: number;
  goal?: string;
  dietaryPreferences?: string[];
}

export interface WellnessSummary {
  proteinIntake: number;
  proteinGoal: number;
  waterIntake: number;
  waterGoal: number;
  caloriesConsumed: number;
  caloriesGoal: number;
  streak: number;
  wellnessScore: number;
  aiInsight: string;
  topMeals: Meal[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type OpenaiImageInputSize = typeof OpenaiImageInputSize[keyof typeof OpenaiImageInputSize];


export const OpenaiImageInputSize = {
  '1024x1024': '1024x1024',
  '512x512': '512x512',
  '256x256': '256x256',
} as const;

export interface OpenaiImageInput {
  prompt: string;
  size?: OpenaiImageInputSize;
}

export interface OpenaiImageOutput {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type ListMealsParams = {
filter?: ListMealsFilter;
search?: string;
};

export type ListMealsFilter = typeof ListMealsFilter[keyof typeof ListMealsFilter];


export const ListMealsFilter = {
  'high-protein': 'high-protein',
  'low-carb': 'low-carb',
  budget: 'budget',
  'diabetic-friendly': 'diabetic-friendly',
  'gym-meals': 'gym-meals',
} as const;