# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NutriFlow AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: meals
    description: Meal recommendation operations
  - name: restaurants
    description: Restaurant operations
  - name: grocery
    description: Grocery planning operations
  - name: profile
    description: User profile and wellness operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List meal recommendations
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
            enum: [high-protein, low-carb, budget, diabetic-friendly, gym-meals]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of meal recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"

  /meals/{id}:
    get:
      operationId: getMeal
      tags: [meals]
      summary: Get a single meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /meal-plan:
    get:
      operationId: getMealPlan
      tags: [meals]
      summary: Get the current weekly meal plan
      responses:
        "200":
          description: Weekly meal plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlan"

  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List healthy restaurants
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"

  /grocery/list:
    get:
      operationId: getGroceryList
      tags: [grocery]
      summary: Get current grocery list
      responses:
        "200":
          description: Current grocery list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"

  /grocery/plan:
    post:
      operationId: createGroceryPlan
      tags: [grocery]
      summary: Generate an AI grocery plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryPlanInput"
      responses:
        "201":
          description: Generated grocery plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"

  /grocery/items/{id}/toggle:
    patch:
      operationId: toggleGroceryItem
      tags: [grocery]
      summary: Toggle a grocery item checked state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated grocery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /wellness/summary:
    get:
      operationId: getWellnessSummary
      tags: [profile]
      summary: Get wellness dashboard summary
      responses:
        "200":
          description: Wellness summary with nutrition, streaks, and AI insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WellnessSummary"

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

  /openai/conversations/{id}/voice-messages:
    post:
      operationId: sendOpenaiVoiceMessage
      tags: [openai]
      summary: Send audio and receive a streaming voice response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiVoiceMessageInput"
      responses:
        "200":
          description: SSE stream of transcript and audio chunks
          content:
            text/event-stream: {}

  /openai/generate-image:
    post:
      operationId: generateOpenaiImage
      tags: [openai]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiImageOutput"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Meal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        healthScore:
          type: number
        cuisine:
          type: string
        tags:
          type: array
          items:
            type: string
        price:
          type: number
        isAiRecommended:
          type: boolean
        restaurantId:
          type: ["integer", "null"]
      required:
        - id
        - name
        - description
        - calories
        - protein
        - carbs
        - fat
        - healthScore
        - cuisine
        - tags
        - price
        - isAiRecommended

    MealPlanDay:
      type: object
      properties:
        day:
          type: string
        breakfast:
          $ref: "#/components/schemas/Meal"
        lunch:
          $ref: "#/components/schemas/Meal"
        dinner:
          $ref: "#/components/schemas/Meal"
        totalCalories:
          type: integer
        totalProtein:
          type: number
      required:
        - day
        - breakfast
        - lunch
        - dinner
        - totalCalories
        - totalProtein

    MealPlan:
      type: object
      properties:
        weekOf:
          type: string
        days:
          type: array
          items:
            $ref: "#/components/schemas/MealPlanDay"
      required:
        - weekOf
        - days

    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cuisine:
          type: string
        rating:
          type: number
        deliveryTime:
          type: string
        imageUrl:
          type: ["string", "null"]
        isHealthy:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - cuisine
        - rating
        - deliveryTime
        - isHealthy
        - tags

    GroceryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        quantity:
          type: string
        unit:
          type: string
        isChecked:
          type: boolean
        nutritionNote:
          type: ["string", "null"]
      required:
        - id
        - name
        - category
        - quantity
        - unit
        - isChecked

    GroceryList:
      type: object
      properties:
        id:
          type: integer
        weekOf:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroceryItem"
        totalItems:
          type: integer
        checkedItems:
          type: integer
      required:
        - id
        - weekOf
        - items
        - totalItems
        - checkedItems

    GroceryPlanInput:
      type: object
      properties:
        goal:
          type: string
        dietaryPreferences:
          type: array
          items:
            type: string
        budget:
          type: ["number", "null"]
      required:
        - goal

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: ["integer", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        goal:
          type: string
        dietaryPreferences:
          type: array
          items:
            type: string
        wellnessScore:
          type: integer
        streak:
          type: integer
        avatarUrl:
          type: ["string", "null"]
      required:
        - id
        - name
        - goal
        - dietaryPreferences
        - wellnessScore
        - streak

    UserProfileInput:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
        goal:
          type: string
        dietaryPreferences:
          type: array
          items:
            type: string

    WellnessSummary:
      type: object
      properties:
        proteinIntake:
          type: number
        proteinGoal:
          type: number
        waterIntake:
          type: number
        waterGoal:
          type: number
        caloriesConsumed:
          type: integer
        caloriesGoal:
          type: integer
        streak:
          type: integer
        wellnessScore:
          type: integer
        aiInsight:
          type: string
        topMeals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
      required:
        - proteinIntake
        - proteinGoal
        - waterIntake
        - waterGoal
        - caloriesConsumed
        - caloriesGoal
        - streak
        - wellnessScore
        - aiInsight
        - topMeals

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiVoiceMessageInput:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiImageInput:
      type: object
      properties:
        prompt:
          type: string
        size:
          type: string
          enum: ["1024x1024", "512x512", "256x256"]
      required:
        - prompt

    OpenaiImageOutput:
      type: object
      properties:
        b64_json:
          type: string
      required:
        - b64_json

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error