# lib/api-zod/src/generated/types/groceryItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GroceryItem {
  id: number;
  name: string;
  category: string;
  quantity: string;
  unit: string;
  isChecked: boolean;
  /** @nullable */
  nutritionNote?: string | null;
}