# lib/api-zod/src/generated/types/meal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Meal {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  healthScore: number;
  cuisine: string;
  tags: string[];
  price: number;
  isAiRecommended: boolean;
  /** @nullable */
  restaurantId?: number | null;
}