# lib/api-zod/src/generated/types/mealPlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealPlanDay } from './mealPlanDay';

export interface MealPlan {
  weekOf: string;
  days: MealPlanDay[];
}