# lib/api-zod/src/generated/types/mealPlanDay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Meal } from './meal';

export interface MealPlanDay {
  day: string;
  breakfast: Meal;
  lunch: Meal;
  dinner: Meal;
  totalCalories: number;
  totalProtein: number;
}