# lib/api-zod/src/generated/types/openaiVoiceMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}